update to index.js
